# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCL Piloto Classe II API
  version: 0.1.0
paths:
  /api/qr/resolve:
    post:
      summary: Resolve QR MCL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Unidade encontrada
        "400":
          description: Formato inválido
  /api/events:
    post:
      summary: Registra evento logístico
      responses:
        "200":
          description: Evento registrado ou idempotente
        "401":
          description: Autenticação obrigatória
  /api/import:
    post:
      summary: Valida importação CSV/JSON
      responses:
        "200":
          description: Resultado da validação
  /api/connectors/compras-gov/sync:
    post:
      summary: Sincroniza ARP Item do Compras.gov.br
      description: Requer ADMIN ou LOGISTICS_MANAGER. Consulta API publica oficial, registra staging, metricas e auditoria.
      responses:
        "200":
          description: Execucao do conector
        "401":
          description: Autenticacao obrigatoria
        "403":
          description: Permissao insuficiente
        "409":
          description: Sincronizacao concorrente ou falha controlada
  /api/aquisicoes/links:
    post:
      summary: Vincula necessidade a instrumento publico
      description: Cria ObjectLink manual NEED --PODE_SER_ATENDIDA_POR--> ACQUISITION_INSTRUMENT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - needId
                - acquisitionInstrumentId
                - justification
              properties:
                needId:
                  type: string
                acquisitionInstrumentId:
                  type: string
                justification:
                  type: string
                confidence:
                  type: number
                  minimum: 0
                  maximum: 1
      responses:
        "200":
          description: Vinculo criado ou idempotente
        "400":
          description: Entrada invalida
        "401":
          description: Autenticacao obrigatoria
  /api/v1/material-analyses:
    get:
      summary: Lista todas as analises de cobertura de materiais
      responses:
        "200":
          description: Lista de analises
        "401":
          description: Autenticacao obrigatoria
    post:
      summary: Cria ou busca uma analise de cobertura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - needId
              properties:
                needId:
                  type: string
      responses:
        "200":
          description: Analise retornada
        "400":
          description: Parametro invalido
        "401":
          description: Autenticacao obrigatoria
  /api/v1/material-analyses/{id}:
    get:
      summary: Busca detalhes de uma analise especifica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalhes da analise e candidatos
        "401":
          description: Autenticacao obrigatoria
        "404":
          description: Analise nao localizada